chore(website): extract format layer to helper-core package (ADR-0003)
Move the pure-TS backup format layer out of apps/helper/src/lib into a
private workspace package packages/core (helper-core):

- core moves: compat/ cuplivo/ kelivo/ kelivo-v120/ rikkahub/ migrate/
  report.ts zip.ts; source-only package with explicit subpath exports
  consumed via workspace:*
- core owns the regression net: scripts/e2e-test.ts + compat-test.ts
  run via tsx under pnpm -r test
- helper keeps only UI layer (ui/, view.ts, view.svelte.ts) and now
  imports the pipeline from helper-core
- tsx/jszip/sqlite-wasm deps re-homed to where they are used; helper
  keeps jszip as devDep for type-only imports
- docs: website/docs/adr/0003 records the package boundary
