+++ website/apps/helper/package.json
@@ -7,20 +7,18 @@
     "dev": "vite",
     "build": "vite build",
     "preview": "vite preview",
-    "check": "svelte-check --tsconfig ./tsconfig.json",
-    "test": "tsx scripts/e2e-test.ts && tsx scripts/compat-test.ts"
+    "check": "svelte-check --tsconfig ./tsconfig.json"
   },
   "dependencies": {
-    "@sqlite.org/sqlite-wasm": "3.49.2-build1",
-    "jszip": "^3.10.1"
+    "helper-core": "workspace:*"
   },
   "devDependencies": {
     "@sveltejs/vite-plugin-svelte": "^6.0.0",
     "@tailwindcss/vite": "^4.1.0",
+    "jszip": "^3.10.1",
     "svelte": "^5.0.0",
     "svelte-check": "^4.0.0",
     "tailwindcss": "^4.1.0",
-    "tsx": "^4.23.9",
     "typescript": "^5.5.0",
     "vite": "^7.0.0"
   }
